bot-musica/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
PyNaCl==1.5.0
#wavelink==2.8.2 # Se comenta para usar la versión que soporta Python 3.11
wavelink==3.3.0 # Versión compatible con Python 3.11+
youtube_dl==2021.12.17
spotipy==2.23.0
# mcstatus y mcrcon no son necesarios para el bot de música
# mcstatus==11.0.0
# mcrcon==0.7.0
# PySocks no es necesario si no se usa Fixie para este bot
# PySocks==1.7.1
Flask==3.0.3 # Añadido para el health check de Heroku
watchdog>=3.0.0
requests # Asegurarse que requests esté, ya que es una dependencia común y pudo haberse eliminado por error.
